chore: move `Entry` implementation to separate files
