+++ openraft/src/entry/raft_entry.rs
@@ -8,16 +8,9 @@ use crate::Membership;
 use crate::RaftTypeConfig;
 use crate::base::OptionalFeatures;
 use crate::base::finalized::Final;
+use crate::entry::RaftPayload;
 use crate::type_config::alias::CommittedLeaderIdOf;
 use crate::type_config::alias::LogIdOf;
-
-/// Defines operations on an entry payload.
-pub trait RaftPayload<C>
-where C: RaftTypeConfig
-{
-    /// Return `Some(Membership)` if the entry payload contains a membership payload.
-    fn get_membership(&self) -> Option<Membership<C>>;
-}
 
 /// Defines operations on an entry.
 pub trait RaftEntry<C>
